[redhat] configs: Set CONFIG_HZ=1000 for x86_64
Message-id: <20180419191259.28877-1-herton@redhat.com>
Patchwork-id: 7646
O-Subject: [kernel team] [RHEL8 PATCH v2] [redhat] configs: Set CONFIG_HZ=1000 for x86_64
Bugzilla: 1564274
RH-Acked-by: Waiman Long <longman@redhat.com>
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1564274

Historically we set CONFIG_HZ=1000 on x86_64, except on Pegas/RHEL-ALT, where
we defaulted to HZ=100 for testing. Since for RHEL8 we copied the same configs
from RHEL-ALT, we inherited the default setting. However, x86_64 was never
supported on RHEL-ALT, and for the time being we still need HZ=1000 for
networking, as David Miller explained on his email to our internal kernel list:

"When the system is busy, a system with NO_HZ + HZ=1000 processes it's timer
workload more smoothly.

With a lower value of HZ such as 250 or 100, lots of timer events get
"bunched up" due to the lower granularity and thus the timers are
processed in a very bursty fashion.

This hurts TCP with lots of connections on the system."

While the networking concerns aren't addressed, and lower HZ settings
are stil being evaluated, just change back x86_64 setting to 1000.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>