[redhat] use rhpkg with sha512 support
Patchwork-id: 7901
O-Subject: [kernel team] [RHEL8 PATCH] [redhat] use rhpkg with sha512 support
Bugzilla:
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=16269563

Since RHEL 8 enabled and switched to use SHA512 for dist-git
sources, we should also do it for kernel sources we upload.

A new binary is provided now by rhpkg with sha512 (rhpkg-sha512),
which now we need to use if we want to upload sources using sha512.

Change our scripts to use it, and also make it simpler for us to switch
back later to the standard rhpkg binary, if/when it starts to support sha512 by
default (right now it defaults to md5).

I also changed the code in scripts to upload all sources at once and
clean the .gitignore and sources files before uploading, because you
can't upload a source while a source with another hash is present in
.gitignore/sources (rhpkg throws an error if you do that). Also the
format of hash present in the sources files changes with rhpkg-sha512,
so we can't anymore use the previous awk expression.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>