[powerpc] powerpc/pkeys: key allocation/deallocation must not change pkey registers
Message-id: <20180804094910.23122.46249.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 226107
O-Subject: [PATCH RHEL8.0 BZ1498799 3/9] powerpc/pkeys: key allocation/deallocation must not change pkey registers
Bugzilla: 1498799
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1498799

Description:
------------
Key allocation and deallocation has the side effect of programming the
UAMOR/AMR/IAMR registers. This is wrong, since its the responsibility of
the application and not that of the kernel, to modify the permission on
the key.

Do not modify the pkey registers at key allocation/deallocation.

This patch also fixes a bug where a sys_pkey_free() resets the UAMOR
bits of the key, thus making its permissions unmodifiable from user
space. Later if the same key gets reallocated from a different thread
this thread will no longer be able to change the permissions on the key.

Fixes: cf43d3b26452 ("powerpc: Enable pkey subsystem")
Cc: stable@vger.kernel.org # v4.16+
Reviewed-by: Thiago Jung Bauermann <bauerman@linux.ibm.com>
Signed-off-by: Ram Pai <linuxram@us.ibm.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>

Upstream Status:
----------------
commit 4a4a5e5d2aadc793be95024f454cf511d115b62d

https://git.kernel.org/pub/scm/linux/kernel/git/powerpc/linux.git/commit/?h=next&id=4a4a5e5d2aadc793be95024f454cf511d115b62d

----------------
Steve Best

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>