[powerpc] powerpc/pkeys: Preallocate execute-only key
Message-id: <20180804094940.23122.91026.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 226110
O-Subject: [PATCH RHEL8.0 BZ1498799 6/9] powerpc/pkeys: Preallocate execute-only key
Bugzilla: 1498799
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1498799

Description:
------------
execute-only key is allocated dynamically. This is a problem. When a
thread implicitly creates an execute-only key, and resets the UAMOR
for that key, the UAMOR value does not percolate to all the other
threads. Any other thread may ignorantly change the permissions on the
key. This can cause the key to be not execute-only for that thread.

Preallocate the execute-only key and ensure that no thread can change
the permission of the key, by resetting the corresponding bit in
UAMOR.

Fixes: 5586cf61e108 ("powerpc: introduce execute-only pkey")
Cc: stable@vger.kernel.org # v4.16+
Signed-off-by: Ram Pai <linuxram@us.ibm.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>

Upstream Status:
----------------
commit a4fcc877d4e18b5efe26e93f08f0cfd4e278c7d9

https://git.kernel.org/pub/scm/linux/kernel/git/powerpc/linux.git/commit/?h=next&id=a4fcc877d4e18b5efe26e93f08f0cfd4e278c7d9

----------------
Steve Best

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>