[scsi] csiostor: Add a soft dep on cxgb4 driver
Message-id: <1533723737-30548-3-git-send-email-avynipad@redhat.com>
Patchwork-id: 226320
O-Subject: [RHEL8.0 PATCH BZ 1503574 2/3] csiostor: Add a soft dep on cxgb4 driver
Bugzilla: 1503574
RH-Acked-by: Jonathan Toppins <jtoppins@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Ewan Milne <emilne@redhat.com>

Bugzilla: http://bugzilla.redhat.com/1503574
Upstream info: RHEL only

We are facing a couple of issues, when csiostor driver takes control of
Chelsio adapter initialization rather than cxgb4. To be on the safe
side, making csiostor dependent on cxgb4 driver untill those issues
are fixed upstream.
Since the problem is not limited to a single issue, solving this in
upstream will take some time. We will remove this soft dependency from
RHEL once we are sure that all the upstream issues are resolved. We
don't want our customer's basic Networking functionality to be affected
because of this driver(csiostor) issues, and hence the softdep fix.

Cc: Arjun Vynipadath <avynipad@redhat.com>
Cc: Don Dutile <ddutile@redhat.com>
Cc: Jonathan Toppins <jtoppins@redhat.com>
Cc: Maurizio Lombardi <mlombard@redhat.com>
Cc: Neil Horman <nhorman@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>