[virt] kvm: selftests: add basic test for state save and restore
Message-id: <20180814173414.4577-16-pbonzini@redhat.com>
Patchwork-id: 226808
O-Subject: [PATCH RHEL8 RHEL8 15/79] kvm: selftests: add basic test for state save and restore
Bugzilla: 1614808
RH-Acked-by: Bandan Das <bsd@redhat.com>
RH-Acked-by: Vitaly Kuznetsov <vkuznets@redhat.com>

The test calls KVM_RUN repeatedly, and creates an entirely new VM with the
old memory and vCPU state on every exit to userspace.  The kvm_util API is
expanded with two functions that manage the lifetime of a kvm_vm struct:
the first closes the file descriptors and leaves the memory allocated,
and the second opens the file descriptors and reuses the memory from
the previous incarnation of the kvm_vm struct.

For now the test is very basic, as it does not test for example XSAVE or
vCPU events.  However, it will test nested virtualization state starting
with the next patch.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit f1764d4915c9a8702223760f0e7f57c992ff3de7)
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>