[x86] x86/cpufeatures: Increase NCAPINTS for future extension
Message-id: <1536596039-6620-2-git-send-email-longman@redhat.com>
Patchwork-id: 228075
O-Subject: [RHEL8 PATCH v3 1/6] x86/cpufeatures: Increase NCAPINTS for future extension
Bugzilla: 1565180
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1565180
Upstream Status: RHEL-only

Over time in the kernel development process, the number of words used
for x86 CPU feature bits keep increasing. For the 2 words that are used
internally by Linux (words 3 & 7), only 1 bit is now left unused. So it
is likely the current NCAPINTS value (19) will be increased in the near
future.

This patch adds 4 more words to NCAPINTS for future extension, the last
of which is for RHEL specific use. The advantage of pre-allocation is
that we don't need to to add complicated logic and potentally slowing
down feature bit check in case we really need more.

Signed-off-by: Waiman Long <longman@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>