[tools] perf python: Fix pyrf_evlist__read_on_cpu() interface
Message-id: <20180913114949.9847-3-jolsa@redhat.com>
Patchwork-id: 228805
O-Subject: [PATCH RHEL8.0 BZ1628229 2/2] perf python: Fix pyrf_evlist__read_on_cpu() interface
Bugzilla: 1628229
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1628229
https://bugzilla.redhat.com/show_bug.cgi?id=1628229

upstream
========
commit 721f0dfc3ce821c6a32820ab63edfb48ed4af075
Author: Jiri Olsa <jolsa@kernel.org>
Date:   Fri Aug 17 13:45:56 2018 +0200

description
===========
Jaroslav reported errors from valgrind over perf python script:

  # echo 0 > /sys/devices/system/cpu/cpu4/online
  # valgrind ./test.py
  ==7524== Memcheck, a memory error detector
  ...
  ==7524== Command: ./test.py
  ==7524==
  pid 7526 exited
  ==7524== Invalid read of size 8
  ==7524==    at 0xCC2C2B3: perf_mmap__read_forward (evlist.c:780)
  ==7524==    by 0xCC2A681: pyrf_evlist__read_on_cpu (python.c:959)
  ...
  ==7524==  Address 0x65c4868 is 16 bytes after a block of size 459,36..
  ==7524==    at 0x4C2B955: calloc (vg_replace_malloc.c:711)
  ==7524==    by 0xCC2F484: zalloc (util.h:35)
  ==7524==    by 0xCC2F484: perf_evlist__alloc_mmap (evlist.c:978)
  ...

The reason for this is in the python interface, that allows a script to
pass arbitrary cpu number, which is then used to access struct
perf_evlist::mmap array. That's obviously wrong and works only when if
all cpus are available and fails if some cpu is missing, like in the
example above.

This patch makes pyrf_evlist__read_on_cpu() search the evlist's maps
array for the proper map to access.

It's linear search at the moment. Based on the way how is the
read_on_cpu used, I don't think we need to be fast in here.  But we
could add some hash in the middle to make it fast/er.

We don't allow python interface to set write_backward event attribute,
so it's safe to check only evlist's mmaps.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>