[virt] KVM: arm/arm64: Clean dcache to PoC when changing PTE due to CoW
Message-id: <20180925152750.24993-32-wei@redhat.com>
Patchwork-id: 229197
O-Subject: [RHEL8 BZ1627474 V2 PATCH 31/34] KVM: arm/arm64: Clean dcache to PoC when changing PTE due to CoW
Bugzilla: 1627474
RH-Acked-by: Mark Salter <msalter@redhat.com>
RH-Acked-by: Auger Eric <eric.auger@redhat.com>
RH-Acked-by: Andrew Jones <drjones@redhat.com>

From: Marc Zyngier <marc.zyngier@arm.com>

When triggering a CoW, we unmap the RO page via an MMU notifier
(invalidate_range_start), and then populate the new PTE using another
one (change_pte). In the meantime, we'll have copied the old page
into the new one.

The problem is that the data for the new page is sitting in the
cache, and should the guest have an uncached mapping to that page
(or its MMU off), following accesses will bypass the cache.

In a way, this is similar to what happens on a translation fault:
We need to clean the page to the PoC before mapping it. So let's just
do that.

This fixes a KVM unit test regression observed on a HiSilicon platform,
and subsequently reproduced on Seattle.

Fixes: a9c0e12ebee5 ("KVM: arm/arm64: Only clean the dcache on translation fault")
Cc: stable@vger.kernel.org # v4.16+
Reported-by: Mike Galbraith <efault@gmx.de>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Christoffer Dall <christoffer.dall@arm.com>
(cherry picked from commit 694556d54f354d3fe43bb2e61fd6103cca2638a4)
Signed-off-by: Wei Huang <wei@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>