[redhat] config: Enable CRC7
Message-id: <20180926031120.bli3ueh3brhduu5k@gondor.apana.org.au>
Patchwork-id: 229298
O-Subject: [RHEL-8.0 PATCH] config: Enable CRC7
Bugzilla: 1589165
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

BZ: 1589165
Brew https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=18437906

CRC7 is selected by MMC_SPI which is enabled on aarch64.  This
creates a conflict because it is currently disabled in the generic
configuration.

As this algorithm helper is safe to enable on all architectures,
this patch does exactly that.

Signed-off-by: Herbert Xu <herbert.xu@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>