[wireless] ath10k: prevent active scans on potential unusable channels
Message-id: <5f75078bfbaf88ee95c7d8df16f4a842d5ef2276.1539094325.git.poros@redhat.com>
Patchwork-id: 230167
O-Subject: [PATCH RHEL8.0 BZ1637528 05/17] ath10k: prevent active scans on potential unusable channels
Bugzilla: 1637528
RH-Acked-by: Stanislaw Gruszka <sgruszka@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

From: Sven Eckelmann <sven.eckelmann@openmesh.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1637528
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=18699688

commit 3f259111583801013cb605bb4414aa529adccf1c
Author:    Sven Eckelmann <sven.eckelmann@openmesh.com>
Date:      Thu Jul 26 15:59:48 2018 +0200

The QCA4019 hw1.0 firmware 10.4-3.2.1-00050 and 10.4-3.5.3-00053 (and most
likely all other) seem to ignore the WMI_CHAN_FLAG_DFS flag during the
scan. This results in transmission (probe requests) on channels which are
not "available" for transmissions.

Since the firmware is closed source and nothing can be done from our side
to fix the problem in it, the driver has to work around this problem. The
WMI_CHAN_FLAG_PASSIVE seems to be interpreted by the firmware to not
scan actively on a channel unless an AP was detected on it. Simple probe
requests will then be transmitted by the STA on the channel.

ath10k must therefore also use this flag when it queues a radar channel for
scanning. This should reduce the chance of an active scan when the channel
might be "unusable" for transmissions.

Fixes: e8a50f8ba44b ("ath10k: introduce DFS implementation")
Signed-off-by: Sven Eckelmann <sven.eckelmann@openmesh.com>
Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>