[powerpc] KVM: PPC: Book3S HV: Handle hypervisor instruction faults better
Message-id: <1539150936-21109-12-git-send-email-sursingh@redhat.com>
Patchwork-id: 230372
O-Subject: [RHEL-8 kernel PATCH 11/34] KVM: PPC: Book3S HV: Handle hypervisor instruction faults better
Bugzilla: 1505999
RH-Acked-by: Laurent Vivier <lvivier@redhat.com>
RH-Acked-by: David Gibson <dgibson@redhat.com>

From: Paul Mackerras <paulus@ozlabs.org>

Currently the code for handling hypervisor instruction page faults
passes 0 for the flags indicating the type of fault, which is OK in
the usual case that the page is not mapped in the partition-scoped
page tables.  However, there are other causes for hypervisor
instruction page faults, such as not being to update a reference
(R) or change (C) bit.  The cause is indicated in bits in HSRR1,
including a bit which indicates that the fault is due to not being
able to write to a page (for example to update an R or C bit).
Not handling these other kinds of faults correctly can lead to a
loop of continual faults without forward progress in the guest.

In order to handle these faults better, this patch constructs a
"DSISR-like" value from the bits which DSISR and SRR1 (for a HISI)
have in common, and passes it to kvmppc_book3s_hv_page_fault() so
that it knows what caused the fault.

Reviewed-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Paul Mackerras <paulus@ozlabs.org>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
(cherry picked from commit 32eb150aee8dce5076638ebafe53732276c9f28e)

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1505999

Signed-off-by: Suraj Jitindar Singh <sursingh@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>