[netdrv] mlxsw: spectrum_switchdev: Do not leak RIFs when removing bridge
Message-id: <20181011201016.32055-1-poros@redhat.com>
Patchwork-id: 230858
O-Subject: [PATCH RHEL8.0 BZ1638268] mlxsw: spectrum_switchdev: Do not leak RIFs when removing bridge
Bugzilla: 1638268
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1638268
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=18744132

commit 602b74eda81311dbdb5dbab08c30f789f648ebdc
Author: Ido Schimmel <idosch@mellanox.com>
Date:   Fri Aug 24 15:41:35 2018 +0300

When a bridge device is removed, the VLANs are flushed from each
configured port. This causes the ports to decrement the reference count
on the associated FIDs (filtering identifier). If the reference count of
a FID is 1 and it has a RIF (router interface), then this RIF is
destroyed.

However, if no port is member in the VLAN for which a RIF exists, then
the RIF will continue to exist after the removal of the bridge. To
reproduce:

The RIF associated with br0.10 continues to exist.

Fix this by iterating over all the bridge device uppers when it is
destroyed and take care of destroying their RIFs.

Fixes: 99f44bb3527b ("mlxsw: spectrum: Enable L3 interfaces on top of bridge devices")
Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Reviewed-by: Petr Machata <petrm@mellanox.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>