[powerpc] powerpc: Detect the presence of big-cores via ibm, thread-groups
Message-id: <20181015160403.35987.79527.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 230946
O-Subject: [PATCH RHEL8.0 BZ1639265 1/3] powerpc: Detect the presence of big-cores via "ibm, thread-groups"
Bugzilla: 1639265
RH-Acked-by: Gopal Tiwari <gtiwari@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1639265

Description:
------------
On IBM POWER9, the device tree exposes a property array identifed by
"ibm,thread-groups" which will indicate which groups of threads share
a particular set of resources.

As of today we only have one form of grouping identifying the group of
threads in the core that share the L1 cache, translation cache and
instruction data flow.

This patch adds helper functions to parse the contents of
"ibm,thread-groups" and populate a per-cpu variable to cache
information about siblings of each CPU that share the L1, traslation
cache and instruction data-flow.

It also defines a new global variable named "has_big_cores" which
indicates if the cores on this configuration have multiple groups of
threads that share L1 cache.

For each online CPU, it maintains a cpu_smallcore_mask, which
indicates the online siblings which share the L1-cache with it.

Signed-off-by: Gautham R. Shenoy <ego@linux.vnet.ibm.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>

Upstream Status:
----------------
commit 425752c63b6f3fed7b5a9cba2b8101a92cf36995

https://git.kernel.org/pub/scm/linux/kernel/git/powerpc/linux.git/commit/?h=next&id=425752c63b6f3fed7b5a9cba2b8101a92cf36995

----------------
Steve Best

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>