[redhat] configs: Cleanup zfcpdump configs
Message-id: <20181116212057.2779-1-dzickus@redhat.com>
Patchwork-id: 233128
O-Subject: [RHEL8 PATCH 04/11 V3] configs: Cleanup zfcpdump configs
Bugzilla: 1649567
RH-Acked-by: Philipp Rudo <prudo@redhat.com>
RH-Acked-by: Cornelia Huck <cohuck@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1649567

The output of the zfcpdump configs did not match the input configs as
noticed by process_configs.sh -c.

This patch modifies the configs to match the output.

It is assumed the output is correct (as that is what is seen and tested today).

Testing was done by running 'make rh-configs' and saving the configs.
This patch was applied, 'make rh-configs' re-ran and saved.

The difference between the old and new configs was none.

Because of the way zfcpdump is configured, this patch is rather larger to
account for the changes in that kernel.

V2: disable CONFIG_CMA and CONFIG_I2c
V3: regenerated patch with better ouptut from process_configs.sh
    - dropped CONFIG_I2C_ALGOBIT
    - dropped CONFIG_IRQ_POLL
    - dropped CONFIG_SERIO
    - dropped CONFIG_SERIO_LIBPS2
    - dropped CONFIG_SERIO_RAW
    - dropped CONFIG_SERIO_SERPORT

Cc: Philipp Rudo <prudo@redhat.com>
Cc: Hendrik Brueckner <brueckner@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>