[redhat] configs: checkoptions has backwards input
Message-id: <20181120034744.15725-1-dzickus@redhat.com>
Patchwork-id: 233185
O-Subject: [RHEL8 PATCH] configs: checkoptions has backwards input
Bugzilla: 1655718
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1655718

The input to checkoptions in process_configs.sh was backwards.  It uses

"diff $new $old" instead of "diff $old $new".

When trying to check if a config option has been modified, the check was
successful, but the output message was confusing.

Found # CONFIG_FOO is not set, after generation, had CONFIG_FOO y in Source tree

instead of

Found CONFIG_FOO=y after generation, had CONFIG_FOO=is not set in Source tree

As a result, when trying to use the output message as a reference for what the
config option should be modified to, it was a struggle.

There are two cases in the awk script: /is not set/ and /=/.  By fixing the
inputs, we have to fix the /=/ case too.  The /is not set/ case was backwards
to begin with and does not need touching.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>