[powerpc] powerpc/powernv/npu: Add a debugfs setting to change ATSD threshold
Message-id: <20181127183842.34118.51715.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 233776
O-Subject: [PATCH RHEL8.0 BZ1653661 1/4] powerpc/powernv/npu: Add a debugfs setting to change ATSD threshold
Bugzilla: 1653661
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Gopal Tiwari <gtiwari@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1653661

Description:
------------
The threshold at which it becomes more efficient to coalesce a range
of ATSDs into a single per-PID ATSD is currently not well understood
due to a lack of real-world work loads. This patch adds a debugfs
parameter allowing the threshold to be altered at runtime in order to
aid future development and refinement of the value.

Signed-off-by: Alistair Popple <alistair@popple.id.au>
Acked-by: Balbir Singh <bsingharora@gmail.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>

Upstream Status:
----------------
commit 99c3ce33a00bc40cb218af770ef00c82c8044c36

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=99c3ce33a00bc40cb218af770ef00c82c8044c36
----------------
Steve Best

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>