[x86] x86/spec_ctrl: Change default Skylake Spectre v2 mitigation to retpoline
Message-id: <1543439786-4127-1-git-send-email-longman@redhat.com>
Patchwork-id: 233890
O-Subject: [RHEL8 PATCH v2] x86/spec_ctrl: Change default Skylake Spectre v2 mitigation to retpoline
Bugzilla: 1651806
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Josh Poimboeuf <jpoimboe@redhat.com>
RH-Acked-by: Andrea Arcangeli <aarcange@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1651806
Upstream: RHEL only
BuildInfo: https://brewweb.devel.redhat.com/taskinfo?taskID=19297614

RHEL7 defaults to use IBRS for Skylake-based CPUs. However, this does
come with a performance overhead which puts RHEL at a performance
disadvantage when compared with other distros that default to retpoline.

According to Intel's document, Skylake-based CPUs will require empty RSB
mitigation (RSB stuffing) when the following asynchronous events happen:

 1. Interrupts/NMIs/traps/aborts/exceptions which increase call depth.
 2. System Management Interrupts (SMI) (see BIOS/Firmware Interactions).
 3. Host VMEXIT/VMRESUME/VMENTER.
 4. Microcode update load (WRMSR 0x79) on another logical processor of
    the same core.

Skylake has a RSB depth 32. If the RSB "stack" becomes empty, a RET
instruction may speculate based on the contents of the indirect branch
predictor, the structure that retpoline is designed to avoid.

However, if a malicious user can find a reliable way to trigger the
kernel to do more than 32 nested function calls (like in a nested VM
reclaim path), it can cause the RSB to become empty in kernel mode
leading the CPU to speculate on the indirect branch predictor.

Perhaps, we should conditionally enable RSB stuffing on NMI/IRQ exit on
Skylake-based CPUs and on return to kernel space so as to reduce
the chance of RSB becoming empty. We will need to do more performance
measurement to find out the cost of doing that.

We don't know if it's possible to find a reproducible path that executes
enough ret instructions in a row to cause an RSB underflow on skylake
CPUs, but for the time being we have the duty to follow the best practice
recommended by the CPU manufacturer for their Skylake (and more recent)
CPUs. If somebody will later find a way to reproducibly execute enough
ret instructions in a row in the kernel, the liability is on the CPU
manufacturer side for publically recommending to use the retpoline
mitigation on those CPUs.

RHEL8 performs RSB stuffing on vmexit and context switch. For
Interrupts/NMIs, it is probably assumed that the handler code won't
increase call depth. We may have limited control on SMI, but the
frequency that one can do that doesn't make that path that useful to
a malicious user. Microcode update load is a rare occurrence.

Intel considers the risk of an attack based on exploiting deep call
stacks low. As discussed above, we can't rule out the possibility that
someone may find a way to do that. So a warning message shown below
will be printed to warn users that IBRS should be used for systems
with high security requirement when no spectre_v2 boot commmand line
parameter is given.

 [...] Spectre V2 : Using retpoline on Skylake-generation processors may not fully mitigate the vulnerability.
 [...] Spectre V2 : Add the "spectre_v2=ibrs" kernel boot flag to enable IBRS on Skylake systems that need full mitigation.

We will no longer report the use of retpoline on Skylake+ CPUs as
vulnerable in the sysfs vulnerabilities/spectre_v2 file.

 v2: Update the warning text (suggested by Joe) and the commit log
     (partly from Andrea).

Reviewed-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Waiman Long <longman@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>