[netdrv] ice: use [sr]q.count when checking if queue is initialized
Message-id: <c04e6ce0d426882d6fbbbdfbb9072ba3fe89ae9a.1543611099.git.jtoppins@redhat.com>
Patchwork-id: 234224
O-Subject: [RHEL8.0 PATCH 02/59] ice: use [sr]q.count when checking if queue is initialized
Bugzilla: 1644403
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Corinna Vinschen <vinschen@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

When shutting down the controlqs, we check if they are initialized
before we shut them down and destroy the lock. This is important, as it
prevents attempts to access the lock of an already shutdown queue.

Unfortunately, we checked rq.head and sq.head as the value to determine
if the queue was initialized. This doesn't work, because head is not
reset when the queue is shutdown. In some flows, the adminq will have
already been shut down prior to calling ice_shutdown_all_ctrlqs. This
can result in a crash due to attempting to access the already destroyed
mutex.

Fix this by using rq.count and sq.count instead. Indeed, ice_shutdown_sq
and ice_shutdown_rq already indicate that this is the value we should be
using to determine of the queue was initialized.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Anirudh Venkataramanan <anirudh.venkataramanan@intel.com>
Tested-by: Andrew Bowers <andrewx.bowers@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
(cherry picked from commit c913b73cd05d8167433846eeb5b1d97dc820d7df)
Bugzilla: 1644403
Build Info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=19333075
Tested: tested by intel as no hw in house
Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>