[netdrv] ice: fix changing of ring descriptor size (ethtool -G)
Message-id: <d5bd4200986056dae2446b861282997377d96e90.1543611099.git.jtoppins@redhat.com>
Patchwork-id: 234235
O-Subject: [RHEL8.0 PATCH 09/59] ice: fix changing of ring descriptor size (ethtool -G)
Bugzilla: 1644403
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Corinna Vinschen <vinschen@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

rx_mini_pending was set to an incorrect value. This was causing EINVAL to
always be returned to 'ethtool -G'. The driver does not support mini or
jumbo rings so the respective settings should be zero.

Also, change the valid range of the number of descriptors in the rings to
make the code simpler and easier for users to understand (this removes the
valid settings of 8 and 16). Add a system log message indicating when the
number is rounded-up from what the user specifies with the 'ethtool -G'
command (i.e. when it is not a multiple of 32), and update the log message
when a user-provided value is out of range to also indicate the stride.

Signed-off-by: Bruce Allan <bruce.w.allan@intel.com>
Signed-off-by: Anirudh Venkataramanan <anirudh.venkataramanan@intel.com>
Tested-by: Andrew Bowers <andrewx.bowers@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
(cherry picked from commit 3b6bf296c44f2b1233b7622c847a6ac5b1aeba0f)
Bugzilla: 1644403
Build Info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=19333075
Tested: tested by intel as no hw in house
Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>