[tty] tty: wipe buffer if not echoing data
Message-id: <1544010923-11255-2-git-send-email-crecklin@redhat.com>
Patchwork-id: 234527
O-Subject: [RHEL8.1 BZ1655051 1/2] tty: wipe buffer if not echoing data
Bugzilla: 1655051
RH-Acked-by: Joe Lawrence <joe.lawrence@redhat.com>
RH-Acked-by: Artem Savkov <asavkov@redhat.com>

commit b97b3d9fb57860a60592859e332de7759fd54c2e
Author: Greg KH <greg@kroah.com>
Date:   Thu Oct 4 11:06:14 2018 -0700

tty: wipe buffer if not echoing data

If we are not echoing the data to userspace or the console is in icanon
mode, then perhaps it is a "secret" so we should wipe it once we are
done with it.

This mirrors the logic that the audit code has.

Reported-by: aszlig <aszlig@nix.build>
Tested-by: Milan Broz <gmazyland@gmail.com>
Tested-by: Daniel Zatovic <daniel.zatovic@gmail.com>
Tested-by: aszlig <aszlig@nix.build>
Cc: Willy Tarreau <w@1wt.eu>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>