[netdrv] qed: Prevent a possible deadlock during driver load and unload
Message-id: <1544648827-15804-4-git-send-email-cdupuis@redhat.com>
Patchwork-id: 235218
O-Subject: [PATCH RHEL 7.7 3/4] qed: Prevent a possible deadlock during driver load and unload
Bugzilla: 1652417
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Kamal Heib <kheib@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Tomer Tayar <Tomer.Tayar@cavium.com>

Bugzilla 1652417

commit eaa50fc59e5841910987e90b0438b2643041f508
https://kernel.googlesource.com/pub/scm/linux/kernel/git/torvalds/linux.git

The MFW manages an internal lock to prevent concurrent hardware
(de)initialization of different PFs.
This, together with the busy-waiting for the MFW's responses for commands,
might lead to a deadlock during concurrent load or unload of PFs.
This patch adds the option to sleep within the busy-waiting, and uses it
for the (un)load requests (which are not sent from an interrupt context) to
prevent the possible deadlock.

Signed-off-by: Tomer Tayar <Tomer.Tayar@cavium.com>
Signed-off-by: Ariel Elior <Ariel.Elior@cavium.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>