[netdrv] qed: Avoid sending mailbox commands when MFW is not responsive
Message-id: <1544648827-15804-5-git-send-email-cdupuis@redhat.com>
Patchwork-id: 235220
O-Subject: [PATCH RHEL 7.7 4/4] qed: Avoid sending mailbox commands when MFW is not responsive
Bugzilla: 1652417
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Kamal Heib <kheib@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Tomer Tayar <Tomer.Tayar@cavium.com>

Bugzilla 1652417

commit b310974e041913231b6e3d5d475d4df55c312301
https://kernel.googlesource.com/pub/scm/linux/kernel/git/torvalds/linux.git

Keep sending mailbox commands to the MFW when it is not responsive ends up
with a redundant amount of timeout expiries.
This patch prints the MCP status on the first command which is not
responded, and blocks the following commands.
Since the (un)load request commands might be not responded due to other
PFs, the patch also adds the option to skip the blocking upon a failure.

Signed-off-by: Tomer Tayar <Tomer.Tayar@cavium.com>
Signed-off-by: Ariel Elior <Ariel.Elior@cavium.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>