[redhat] configs: Remove zfcpdump's CONFIG_NVME_CORE=n
Message-id: <20181218214336.10612-1-dzickus@redhat.com>
Patchwork-id: 236107
O-Subject: [RHEL8 PATCH] configs: Remove zfcpdump's CONFIG_NVME_CORE=n
Bugzilla: 1660639
RH-Acked-by: Hendrik Brueckner <brueckner@redhat.com>
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Cornelia Huck <cohuck@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1660639
Tested: ran 'make rh-configs', no difference to previous configs
Brew: none

It was noticed during verification of bz 1649564 that I missed a config that
set =n.  It was missed because it was never in the final output so the
process_configs.sh -c option never tripped over it.  It had to be manually
found.

I grep'd through the tree for more '=n' options and none were found.

This option is not in the final zfcpdump config file, so instead of
converting it to 'is not set', just remove it.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>