[hv] vmbus: fix subchannel removal
Message-id: <1545231100-15942-12-git-send-email-mgamal@redhat.com>
Patchwork-id: 236178
O-Subject: [RHEL8 KERNEL PATCH v3 11/13] vmbus: fix subchannel removal
Bugzilla: 1650149
RH-Acked-by: Cathy Avery <cavery@redhat.com>
RH-Acked-by: Vitaly Kuznetsov <vkuznets@redhat.com>

The changes to split ring allocation from open/close, broke
the cleanup of subchannels. This resulted in problems using
uio on network devices because the subchannel was left behind
when the network device was unbound.

The cause was in the disconnect logic which used list splice
to move the subchannel list into a local variable. This won't
work because the subchannel list is needed later during the
process of the rescind messages (relid2channel).

The fix is to just leave the subchannel list in place
which is what the original code did. The list is cleaned
up later when the host rescind is processed.

Fixes: ae6935ed7d42 ("vmbus: split ring buffer allocation from open")
Signed-off-by: Stephen Hemminger <sthemmin@microsoft.com>

RHEL-only
Sent on mailing list. Not yet upstream.
See: https://www.spinics.net/lists/linux-driver-devel/msg120802.html

Signed-off-by: Mohammed Gamal <mgamal@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>