[net] redhat: blacklist auto-loadable net modules in modules-extra
Message-id: <950cf24e3005b0ebc39417fd81e5325f1dc86cc8.1546954304.git.mleitner@redhat.com>
Patchwork-id: 236918
O-Subject: [RHEL8.0 net 2/2 v2] redhat: blacklist autoloadable net modules in modules-extra
Bugzilla: 1642795
RH-Acked-by: Stefano Brivio <sbrivio@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1642795
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=19697494

Modules in modules-extra are non-standard modules, which the sysadmin
may not be fully aware of their purpose and how they are used.

ProdSec wants to reduce the attack surface on the system. That, in the
context of above, means not auto-loading odd modules just because an
user triggered it: the sysadmin will have to enable it.

One example is sctp. Currently sctp will be loaded as soon as any user
space application request such socket. SCTP, as not being a mainstream
protocol, is not known to quite some sysadmins, which may have not
considered it on their firewalls, for example.

v1->v2:
- Updated comment on modprobe.d files with Jamie's text, including
  the knowledgebase link for more information.
- s/auto-loadable/autoloadable/ from Stefano
- Safer path/suffix removal, from Stefano/Jiri

Signed-off-by: Marcelo Ricardo Leitner <mleitner@redhat.com>
Signed-off-by: Timothy Redaelli <tredaelli@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>