[redhat] configs: Set defaults for numa configs
Message-id: <20190109184950.3909-20-dzickus@redhat.com>
Patchwork-id: 237063
O-Subject: [RHEL8 PATCH 19/21] configs: Set defaults for numa configs
Bugzilla: 1664813
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Phil Auld <pauld@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Upstream Status: RHEL only.
Bugzilla: https://bugzilla.redhat.com/1664813

Hi,

As part of the effort to establish a baseline for default configs, the
following configs need to reviewed and defined.  Please review the changes
below to see if the defaults are acceptable.  If so, please ACK them, otherwise
NACK with a better default.

Cc: Rafael Aquini <raquini@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>