[redhat] configs: Set defaults for fs configs
Message-id: <20190109184950.3909-21-dzickus@redhat.com>
Patchwork-id: 237062
O-Subject: [RHEL8 PATCH 20/21] configs: Set defaults for fs configs
Bugzilla: 1664813
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Ilya Dryomov <idryomov@redhat.com>
RH-Acked-by: Eric Sandeen <esandeen@redhat.com>

Upstream Status: RHEL only.
Bugzilla: https://bugzilla.redhat.com/1664813

Hi,

As part of the effort to establish a baseline for default configs, the
following configs need to reviewed and defined.  Please review the changes
below to see if the defaults are acceptable.  If so, please ACK them, otherwise
NACK with a better default.

Cc: Eric Sandeen <sandeen@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>