[kvm] kvm: arm64: Add helper for loading the stage2 setting for a VM
Message-id: <1547117027-20773-4-git-send-email-crecklin@redhat.com>
Patchwork-id: 237174
O-Subject: [RHEL8.0 BZ1643586 v5 03/20] kvm: arm64: Add helper for loading the stage2 setting for a VM
Bugzilla: 1643586 1643522
RH-Acked-by: Auger Eric <eric.auger@redhat.com>
RH-Acked-by: Andrew Jones <drjones@redhat.com>

From: Suzuki K Poulose <suzuki.poulose@arm.com>

kvm: arm64: Add helper for loading the stage2 setting for a VM

commit 9f98ddd6686cc9469fb73b11ddd403271d65cbdf
Author: Suzuki K Poulose <suzuki.poulose@arm.com>
Date:   Wed Sep 26 17:32:39 2018 +0100

kvm: arm64: Add helper for loading the stage2 setting for a VM

We load the stage2 context of a guest for different operations,
including running the guest and tlb maintenance on behalf of the
guest. As of now only the vttbr is private to the guest, but this
is about to change with IPA per VM. Add a helper to load the stage2
configuration for a VM, which could do the right thing with the
future changes.

Cc: Christoffer Dall <cdall@kernel.org>
Cc: Marc Zyngier <marc.zyngier@arm.com>
Reviewed-by: Eric Auger <eric.auger@redhat.com>
Signed-off-by: Suzuki K Poulose <suzuki.poulose@arm.com>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>