[kvm] kvm: arm64: Allow tuning the physical address size for VM
Message-id: <1547117027-20773-18-git-send-email-crecklin@redhat.com>
Patchwork-id: 237170
O-Subject: [RHEL8.0 BZ1643586 v5 17/20] kvm: arm64: Allow tuning the physical address size for VM
Bugzilla: 1643586 1643522
RH-Acked-by: Auger Eric <eric.auger@redhat.com>
RH-Acked-by: Andrew Jones <drjones@redhat.com>

RHEL8 note -
      in include/uapi/linux/kvm.h KVM_CAP_ARM_VM_IPA_SIZE set to 165 per
      upstream merge commit e42b4a507efa19a90c63e7968c93c4f82d3bc805

commit 233a7cb235318223df8133235383f4c595c654c1
Author: Suzuki K Poulose <suzuki.poulose@arm.com>
Date:   Wed Sep 26 17:32:54 2018 +0100

Allow specifying the physical address size limit for a new
VM via the kvm_type argument for the KVM_CREATE_VM ioctl. This
allows us to finalise the stage2 page table as early as possible
and hence perform the right checks on the memory slots
without complication. The size is encoded as Log2(PA_Size) in
bits[7:0] of the type field. For backward compatibility the
value 0 is reserved and implies 40bits. Also, lift the limit
of the IPA to host limit and allow lower IPA sizes (e.g, 32).

The userspace could check the extension KVM_CAP_ARM_VM_IPA_SIZE
for the availability of this feature. The cap check returns the
maximum limit for the physical address shift supported by the host.

Cc: Marc Zyngier <marc.zyngier@arm.com>
Cc: Christoffer Dall <cdall@kernel.org>
Cc: Peter Maydell <peter.maydell@linaro.org>
Cc: Paolo Bonzini <pbonzini@redhat.com>
Cc: Radim Krčmář <rkrcmar@redhat.com>
Reviewed-by: Eric Auger <eric.auger@redhat.com>
Signed-off-by: Suzuki K Poulose <suzuki.poulose@arm.com>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>