[redhat] kabi: Adding symbol yield (kernel/sched/core.c)
Message-id: <20190125121331.16146-55-ckalina@redhat.com>
Patchwork-id: 238411
O-Subject: [RHEL8.0 KABI 54/70] kabi: Adding symbol yield (kernel/sched/core.c)
Bugzilla: 1645821 1548617
RH-Acked-by: Josh Poimboeuf <jpoimboe@redhat.com>
RH-Acked-by: Phil Auld <pauld@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1548617,1645821
https://bugzilla.redhat.com/show_bug.cgi?id=1548617
https://bugzilla.redhat.com/show_bug.cgi?id=1645821

Description:
============
Justification not provided for Phase A symbol review.

Type dependency graph:
======================
Note: This section is generated based on the .symtypes file
corresponding to the source file where this symbol is defined. It lists
all the types reachable form this symbol without all the types which are
already reachable from other already whitelisted symbol. Thus it can be
used as a reference of the additional types which will be frozen should
this symbol be accepted for the kABI whitelist.

void yield ( void )

Cc: Herton R. Krzesinski <herton@redhat.com>
Signed-off-by: Čestmír Kalina <ckalina@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>