[redhat] kabi: Adding symbol _ctype (lib/ctype.c)
Message-id: <20190125121854.16283-118-ckalina@redhat.com>
Patchwork-id: 238718
O-Subject: [RHEL8.0 KABI 117/179] kabi: Adding symbol _ctype (lib/ctype.c)
Bugzilla: 1650991 1644033 1643596
RH-Acked-by: Josh Poimboeuf <jpoimboe@redhat.com>
RH-Acked-by: Phil Auld <pauld@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1643596,1644033,1650991
https://bugzilla.redhat.com/show_bug.cgi?id=1643596
https://bugzilla.redhat.com/show_bug.cgi?id=1644033
https://bugzilla.redhat.com/show_bug.cgi?id=1650991

Description:
============
Justification not provided for Phase A symbol review.

Type dependency graph:
======================
Note: This section is generated based on the .symtypes file
corresponding to the source file where this symbol is defined. It lists
all the types reachable form this symbol without all the types which are
already reachable from other already whitelisted symbol. Thus it can be
used as a reference of the additional types which will be frozen should
this symbol be accepted for the kABI whitelist.

extern const unsigned char _ctype [ ]

Cc: Herton R. Krzesinski <herton@redhat.com>
Signed-off-by: Čestmír Kalina <ckalina@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>