[redhat] kabi: Adding symbol vfree (mm/nommu.c)
Message-id: <20190125121909.16381-36-ckalina@redhat.com>
Patchwork-id: 238528
O-Subject: [RHEL8.0 KABI 35/79] kabi: Adding symbol vfree (mm/nommu.c)
Bugzilla: 1650991 1647185 1645821 1644031 1548617 1503871
RH-Acked-by: Josh Poimboeuf <jpoimboe@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1647185,1503871,1548617,1644031,1645821,1650991
https://bugzilla.redhat.com/show_bug.cgi?id=1647185
https://bugzilla.redhat.com/show_bug.cgi?id=1503871
https://bugzilla.redhat.com/show_bug.cgi?id=1548617
https://bugzilla.redhat.com/show_bug.cgi?id=1644031
https://bugzilla.redhat.com/show_bug.cgi?id=1645821
https://bugzilla.redhat.com/show_bug.cgi?id=1650991

Description:
============
Justification not provided for Phase A symbol review.

Type dependency graph:
======================
Note: This section is generated based on the .symtypes file
corresponding to the source file where this symbol is defined. It lists
all the types reachable form this symbol without all the types which are
already reachable from other already whitelisted symbol. Thus it can be
used as a reference of the additional types which will be frozen should
this symbol be accepted for the kABI whitelist.

extern void vfree ( const void * )

Cc: Herton R. Krzesinski <herton@redhat.com>
Cc: Larry Woodman <lwoodman@redhat.com>
Signed-off-by: Čestmír Kalina <ckalina@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>