[redhat] kabi: Adding symbol kzfree (mm/slab_common.c)
Message-id: <20190125121909.16381-61-ckalina@redhat.com>
Patchwork-id: 238562
O-Subject: [RHEL8.0 KABI 60/79] kabi: Adding symbol kzfree (mm/slab_common.c)
Bugzilla: 1647190
RH-Acked-by: Josh Poimboeuf <jpoimboe@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1647190
https://bugzilla.redhat.com/show_bug.cgi?id=1647190

Description:
============
Justification not provided for Phase A symbol review.

Type dependency graph:
======================
Note: This section is generated based on the .symtypes file
corresponding to the source file where this symbol is defined. It lists
all the types reachable form this symbol without all the types which are
already reachable from other already whitelisted symbol. Thus it can be
used as a reference of the additional types which will be frozen should
this symbol be accepted for the kABI whitelist.

void kzfree ( const void * )

Cc: Herton R. Krzesinski <herton@redhat.com>
Cc: Larry Woodman <lwoodman@redhat.com>
Signed-off-by: Čestmír Kalina <ckalina@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>