[redhat] kabi: Adding symbol mem_section (mm/sparse.c)
Message-id: <20190125121909.16381-69-ckalina@redhat.com>
Patchwork-id: 238613
O-Subject: [RHEL8.0 KABI 68/79] kabi: Adding symbol mem_section (mm/sparse.c)
Bugzilla: 1645192 1644036 1644034
RH-Acked-by: Josh Poimboeuf <jpoimboe@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1644034,1645192,1644036
https://bugzilla.redhat.com/show_bug.cgi?id=1644034
https://bugzilla.redhat.com/show_bug.cgi?id=1645192
https://bugzilla.redhat.com/show_bug.cgi?id=1644036

Description:
============
Justification not provided for Phase A symbol review.

Type dependency graph:
======================
Note: This section is generated based on the .symtypes file
corresponding to the source file where this symbol is defined. It lists
all the types reachable form this symbol without all the types which are
already reachable from other already whitelisted symbol. Thus it can be
used as a reference of the additional types which will be frozen should
this symbol be accepted for the kABI whitelist.

struct mem_section { ... }

Cc: Herton R. Krzesinski <herton@redhat.com>
Cc: Larry Woodman <lwoodman@redhat.com>
Signed-off-by: Čestmír Kalina <ckalina@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>