[block] deprecate elevator= kernel parameter
Message-id: <20190125221926.863-3-jmoyer@redhat.com>
Patchwork-id: 238865
O-Subject: [RHEL8.0 PATCH 2/2] block: deprecate elevator= kernel parameter
Bugzilla: 1665295
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Gopal Tiwari <gtiwari@redhat.com>

BZ: 1665295
Brew: 19956180
Upstream Status: RHEL-only

Upstream got rid of this.  Our support team is worried that if we don't
continue to support the option, we will get a bunch of support calls.
So, we added a udev rule, but we'd like to get rid of that in future
major releases.  This patch prints a deprecation warning on boot if
elevator= is specified.

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>