[redhat] kabi: Adding symbol match_token (lib/parser.c)
Message-id: <20190130225056.5455-47-ckalina@redhat.com>
Patchwork-id: 239694
O-Subject: [RHEL8.0 KABI v3 46/58] kabi: Adding symbol match_token (lib/parser.c)
Bugzilla: 1647190 1643321
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1647190,1643321
https://bugzilla.redhat.com/show_bug.cgi?id=1647190
https://bugzilla.redhat.com/show_bug.cgi?id=1643321

Description:
============
Justification not provided for Phase A symbol review.

Type dependency graph:
======================
Note: This section is generated based on the .symtypes file
corresponding to the source file where this symbol is defined. It lists
all the types reachable form this symbol without all the types which are
already reachable from other already whitelisted symbol. Thus it can be
used as a reference of the additional types which will be frozen should
this symbol be accepted for the kABI whitelist.

struct match_token { ... }

Cc: Herton R. Krzesinski <herton@redhat.com>
Signed-off-by: Čestmír Kalina <ckalina@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>