[redhat] configs: Fix missing configs using lowercase letters in rh-configs output
Message-id: <20190207174946.31526-1-dzickus@redhat.com>
Patchwork-id: 241157
O-Subject: [RHEL8 PATCH 1/1] configs: Fix missing configs using lowercase letters in rh-configs output
Bugzilla:
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Herton R. Krzesinski <herton@redhat.com>

Bugzilla: Internal
Brew: Internal
Upstream: N/A
Tested: 'make rh-configs-prep' and look for CONFIG_SCSI_DC395x and
        CONFIG_SYSTEM_TRUSTED_KEYS (process_configs.sh -n exposes this)

Note: This was added to ARK awhile ago but never made it to RHEL-8.

A recent RHEL8 patch cfa4964e22ea2 attempted to fix an issue with
CONFIG_SYSTEM_TRUSTED_KEYS not being set correctly.

The problem with that config option was merge.pl was filtering on the '=' and
the config file used it in a comment that overrode the actual value.  So it
never showed up in the final config.

The above patch tried to fix it by adding a 'grep' command to build_configs.sh
to filter out comments.

That patch falsely assumed all config options used uppercase letters.  This led
to the option 'CONFIG_SCSI_DC395x' being filtered out because it had a
lowercase 'x'.

The are two ways to solve this.  The easy one is to add lowercase filtering to
the grep command in build_configs.sh.  The other is to add some comment
filtering to merge.pl.

This patch chooses the second solution because I was afraid more exotic config
option changes might show up causing us to add more symbols to the grep
expression.  Further the perl script is already reading line-by-line, adding
a simple filter for a comment is easy.

The end result is a partial revert of cfa4964e22ea2 and a different solution.

Cc: Jakub Racek <jracek@redhat.com>
Cc: Herton R. Krzesinski <herton@redhat.com>
Cc: Don Zickus <dzickus@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>