[net] bpf: restrict unknown scalars of mixed signed bounds for unprivileged
Message-id: <20190212111451.15906-4-jolsa@redhat.com>
Patchwork-id: 241704
O-Subject: [PATCH RHEL8 BZ1673631 3/7] bpf: restrict unknown scalars of mixed signed bounds for unprivileged
Bugzilla: 1673631
CVE: CVE-2019-7308
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1673631
https://bugzilla.redhat.com/show_bug.cgi?id=1673631

upstream
========
commit 9d7eceede769f90b66cfa06ad5b357140d5141ed
Author: Daniel Borkmann <daniel@iogearbox.net>
Date: Thu Jan 3 00:58:32 2019 +0100

description
===========
For unknown scalars of mixed signed bounds, meaning their smin_value is
negative and their smax_value is positive, we need to reject arithmetic
with pointer to map value. For unprivileged the goal is to mask every
map pointer arithmetic and this cannot reliably be done when it is
unknown at verification time whether the scalar value is negative or
positive. Given this is a corner case, the likelihood of breaking should
be very small.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>