[redhat] RHELPatchRules: Guidelines for RHEL patch review
Message-id: <alpine.LFD.2.21.1903272216420.3730@clovehitch>
Patchwork-id: 9442
O-Subject: [kernel team] [RHEL8.1 PATCH v4] [redhat] RHELPatchRules: Guidelines for RHEL patch review
Bugzilla:
RH-Acked-by: Marcelo Ricardo Leitner <mleitner@redhat.com>
RH-Acked-by: Stefano Brivio <sbrivio@redhat.com>
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Luis Claudio R. Goncalves <lgoncalv@redhat.com>

The patch volume of rhel kernel updates has grown tremendously.  In
rhel5, the average minor update contained ~1200 patches.  In rhel6
that average grew to ~5200, and in rhel7 it has reached ~13,500.
Obviously, as patch volume continues on this upward trend, it gets
more and more difficult for the kernel maintainer to stay on top of
review discussions.

The current maintainer workflow requires manual review of each and
every patch-reply message to determine if the reply is an ack, nack,
or some other blocking condition.  Additional scenarios such "Ack
conditional on XYZ", partial acking of patchsets, and other
indeterminate patch-replies can further complicate things.  This
frequently requires the maintainer to participate in the review,
rather than simply monitoring it.  This is one of the biggest
time-sinks in the maintainer's workflow.

The kernel maintainers are asking for some changes here, in order to
keep up!

Proposed Solution:

The goal of this patch is to define a clear, concise set of tags
together with their use cases that will allow maintainers to use
automation to assist their workflow.  A secondary benefit is more
adroit review interaction, and hopefully some reduction in email
volume for all participants.

The proposal is to formalize the use of ack/nack tags as follows.

[v4: RFC->PATCH; Add corrections and suggestions from plougher, mleitner, jbenc]

Signed-off-by: Don Howard <dhoward@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>