[infiniband] RDMA/uverbs: Add UVERBS_ATTR_CONST_IN to the specs language
Message-id: <13cfb1fa9920e6749069625e0d1dbfe841488205.1555364289.git.jtoppins@redhat.com>
Patchwork-id: 249981
O-Subject: [PATCH RDMAv4.20 RHEL-8.1 v2 001/139] RDMA/uverbs: Add UVERBS_ATTR_CONST_IN to the specs language
Bugzilla: 1685307
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Kamal Heib <kheib@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

This makes it clear and safe to access constants passed in from user
space. We define a consistent ABI of u64 for all constants, and verify
that the data passed in can be represented by the type the user supplies.

The expectation is this will always be used with an enum declaring the
constant values, and the user will use the enum type as input to the
accessor.

To retrieve the attribute value we introduce two helper calls - one
standard which may fail if attribute is not valid and one where caller can
provide a default value which will be used in case the attribute is not
valid (useful when attribute is optional).

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>
Signed-off-by: Ariel Levkovich <lariel@mellanox.com>
Signed-off-by: Mark Bloch <markb@mellanox.com>
Signed-off-by: Leon Romanovsky <leonro@mellanox.com>
(cherry picked from commit 0953fffec9ba022f63bfe01e86427530d8320d5c)
Bugzilla: 1685307
Build Info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=20935439
Tested: standard RDMA test suite along with partner testing
Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>