[redhat] redhat: enable F29 cross-compiler install for now
Message-id: <20190503223626.19384-2-jarod@redhat.com>
Patchwork-id: 253356
O-Subject: [RHEL8 PATCH 1/2 v3] redhat: enable F29 cross-compiler install for now
Bugzilla:
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Prior to EPEL8 existing, we can use F29's cross-compiler packages to
enable cross builds. The CKI team has been using these for quite some time
now with solid results. Once EPEL-8 is available in a few weeks, that will
still be the preferred method, so this is just a stop-gap, but it may well
be helpful to a few people.

CC: Prarit Bhargava <prarit@redhat.com>
CC: Don Zickus <dzickus@redhat.com>
CC: Major Hayden <mhayden@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>