[redhat] redhat: Make cross-build targets viable
Message-id: <20190503223626.19384-3-jarod@redhat.com>
Patchwork-id: 253357
O-Subject: [RHEL8 PATCH 2/2 v3] redhat: Make cross-build targets viable
Bugzilla:
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Upstream Status: RHEL only

We still need to fix the compile of tools/perf/bpftool and selftests
in a cross environment, but this at least gets the base builds
functional.

This also streamlines the file a bit, but codifying --with cross into
CROSS_RPMFLAGS, so we don't have to repeat it so many times, and in the
-all-rpms case, only do -ba for the last rpm build, since -ba means "do
both -bb and -bs", and we only need -bs (build srpm) once.

v3 of this patch is rebased for kernel-4.18.0-82.el8, which (among other
things) added the kernel-selftests-internal sub-package.

CC: Don Zickus <dzickus@redhat.com>
CC: Major Hayden <mhayden@redhat.com>
CC: Prarit Bhargava <prarit@redhat.com>
Tested-by: Prarit Bhargava <prarit@redhat.com> (v2)
Signed-off-by: Jarod Wilson <jarod@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>