[net] xfrm: Extend the output_mark to support input direction and masking.
Message-id: <E1hNX59-0001Vr-F4@gondobar>
Patchwork-id: 253688
O-Subject: [RHEL8.1 v3 PATCH 1/16] xfrm: Extend the output_mark to support input direction and masking.
Bugzilla: 1702198
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

BZ: 1702198
Brew https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=21527872

Conflicts: Instead of adding xfrm_mark into xfrm_state.props we
instead add a new output_mark_mask field at the end of xfrm_state
in order to preserve kABI.

commit 9b42c1f179a614e11893ae4619f0304a38f481ae

We already support setting an output mark at the xfrm_state,
unfortunately this does not support the input direction and
masking the marks that will be applied to the skb. This change
adds support applying a masked value in both directions.

The existing XFRMA_OUTPUT_MARK number is reused for this purpose
and as it is now bi-directional, it is renamed to XFRMA_SET_MARK.

An additional XFRMA_SET_MARK_MASK attribute is added for setting the
mask. If the attribute mask not provided, it is set to 0xffffffff,
keeping the XFRMA_OUTPUT_MARK existing 'full mask' semantics.

Co-developed-by: Tobias Brunner <tobias@strongswan.org>
Co-developed-by: Eyal Birger <eyal.birger@gmail.com>
Co-developed-by: Lorenzo Colitti <lorenzo@google.com>
Signed-off-by: Steffen Klassert <steffen.klassert@secunet.com>
Signed-off-by: Tobias Brunner <tobias@strongswan.org>
Signed-off-by: Eyal Birger <eyal.birger@gmail.com>
Signed-off-by: Lorenzo Colitti <lorenzo@google.com>
Signed-off-by: Herbert Xu <herbert.xu@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>